backend/requirements.txt
av==0.2.4
gunicorn==19.3.0
pymongo==3.0.3
flask==0.10.1
flask-classy==0.6.10
pyyaml==3.11
requests==2.7.0
scales==1.0.9